+++ packages/tailwindcss-language-server/tests/env/multi-config.test.js
@@ -13,7 +13,7 @@ withFixture('multi-config', (c) => {
       contents: {
         language: 'css',
         value:
-          '.bg-foo {\n  --tw-bg-opacity: 1;\n  background-color: rgb(255 0 0 / var(--tw-bg-opacity));\n}',
+          '& {\n  --tw-bg-opacity: 1;\n  background-color: rgb(255 0 0 / var(--tw-bg-opacity));\n}',
       },
       range: { start: { line: 0, character: 12 }, end: { line: 0, character: 18 } },
     })
@@ -30,7 +30,7 @@ withFixture('multi-config', (c) => {
       contents: {
         language: 'css',
         value:
-          '.bg-foo {\n  --tw-bg-opacity: 1;\n  background-color: rgb(0 0 255 / var(--tw-bg-opacity));\n}',
+          '& {\n  --tw-bg-opacity: 1;\n  background-color: rgb(0 0 255 / var(--tw-bg-opacity));\n}',
       },
       range: { start: { line: 0, character: 12 }, end: { line: 0, character: 18 } },
     })